Portals: Require renderer initiated navigations in portals to be with replacement
For a linear model of portal history, we will only have a portal's last
committed entry upon activation be a part of the user facing session
history. To avoid a portal building up a distinct session history for
itself that serves no purpose, we have renderer initiated navigations in
portals be done with replacement.

For APIs where there is an explicit choice to be made about whether to
replace (e.g. pushState v.s. replaceState), we print a warning if the
option which would create a new entry is used in a portal context.

Bug: 914108
Change-Id: Ib1cdbe3e8d1c0e009eea1f05af28ef2971e8b353